rename pkg name
