+++ pkg/apiserver/diagnose/diagnose_report.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package diagnose_report
+package diagnose
 
 import (
 	"database/sql"
